+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changes
 - Merged changes from JetBrains plugin template
 - [[#120]](https://github.com/Roboroads/laravel-tinker/issues/120) Changes remote tip text in settings
+### Fixed
+- [[#121]](https://github.com/Roboroads/laravel-tinker/issues/121) Exception when the IDE has no foreground color for given attribute
 ## [2.3.0]
 ### Added
 - [[#112]](https://github.com/Roboroads/laravel-tinker/issues/112) Adds <kbd>Open new console</kbd> and <kbd>Reopen last console</kbd> actions to the tinker output toolwindow.